changed errorCode to returnCode and errorMessage to returnMessage. Also added API for retrieving packages.

+++ src/main/java/com/landedexperts/letlock/filetransfer/backend/database/mybatis/response/PaymentResponse.java
@@ -1,14 +1,14 @@
 package com.landedexperts.letlock.filetransfer.backend.database.mybatis.response;
 
 public class PaymentResponse {
-	private final int paymentId;
+	private final long paymentId;
 	private final String message;
 
-	public PaymentResponse(final int paymentId, final String message) {
+	public PaymentResponse(final long paymentId, final String message) {
 		this.paymentId = paymentId;
 		this.message = message;
 	}
 
-	public int getPaymentId() {return this.paymentId; }
+	public long getPaymentId() {return this.paymentId; }
 	public String getMessage() {return this.message;}
 }
